release: :bookmark: version estable index
creation of the folders that will be used during the project

feat: :tada: inicio index.html

feat: :sparkles: footer settings

The footer was created and edited and the necessary images were added for it.

fix: :fire: Header configuration is removed

Several lines of code were eliminated which were responsible for modifying the index header, for future modification

feat: :sparkles: actualizacion de el header (index)

The new form of the head of the index was created, along with its edition in CSS

feat: :sparkles: beginning of balance sheets

information organized into sections was introduced, from the balances of the index

feat: :construction: arrown start

The options (scroll) and size format were added, CSS modification was also added to (arrown), and to the balances (color modifications, spacing, addressing, t sizes, along with the scroll in arrown)

style: :alembic: page update

The CSS format of the complete edition of the index page was modified and new variables were added for better editing

feat: :sparkles: start of transactions

The transactions on Arrown's part were started and due modification was made in css

feat: :sparkles: mycard start and progress

It started and created the footer and header of the mycard page, along with its style configuration.

feat: :sparkles: footer of balance

The configuration of the balance footer was added, with its editing in CSS

fix: :lipstick: virtualWallet responsive

Virtual Wallet arrangement, responsive configuration for all pages

chore: :construction: general details

The old data was updated with the final data in index.html, in the same way the mycard info was styled, giving it the final position and generally colors and margins were modified

fix: :bug: update on mycard

The mycard credit cards were fixed from scratch, along with their configuration, and a code order was made in the same area, where space was cleaned.

chore: :lipstick: font configuration

The fonts were linked to each text in general of the virtualwallet and the font sizes were configured

fix: :art: general arrangements

general arrangements of margin, padding and border radius, also in balance.html an arrangement of images was made

feat: :fire: copia_style_css

creation of the folders that will be used during the project

feat: :fire: index_css

feat: :sparkles: footer settings

The footer was created and edited and the necessary images were added for it.

fix: :fire: Header configuration is removed

Several lines of code were eliminated which were responsible for modifying the index header, for future modification

feat: :sparkles: actualizacion de el header (index)

The new form of the head of the index was created, along with its edition in CSS

feat: :construction: arrown start

The options (scroll) and size format were added, CSS modification was also added to (arrown), and to the balances (color modifications, spacing, addressing, t sizes, along with the scroll in arrown)

style: :alembic: page update

The CSS format of the complete edition of the index page was modified and new variables were added for better editing

feat: :sparkles: start of transactions

The transactions on Arrown's part were started and due modification was made in css

feat: :sparkles: mycard start and progress

It started and created the footer and header of the mycard page, along with its style configuration.

style: :construction: Design Update

Added new total style configuration to ETF, also added some image arrangements in mycard.html

feat: :construction: balance buttons

The buttons in the balance section were created and modified

fix: :hammer: page layout update

configuration is modified and added in balance CSS so that the page begins to be responsive, with which many size and position configurations had to be refactored throughout the page

fix: :lipstick: virtualWallet responsive

Virtual Wallet arrangement, responsive configuration for all pages

chore: :hammer: general configuration

The redirection of the pages was configured, as well as the types of colors and margins in balances and index were modified.

chore: :construction: general details

The old data was updated with the final data in index.html, in the same way the mycard info was styled, giving it the final position and generally colors and margins were modified

fix: :bug: update on mycard

The mycard credit cards were fixed from scratch, along with their configuration, and a code order was made in the same area, where space was cleaned.

chore: :lipstick: font configuration

The fonts were linked to each text in general of the virtualwallet and the font sizes were configured

fix: :art: general arrangements

general arrangements of margin, padding and border radius, also in balance.html an arrangement of images was made