feat: :fire: copia_mycard_html
creation of the folders that will be used during the project

feat: :tada: mycard_html

feat: :sparkles: mycard start and progress

It started and created the footer and header of the mycard page, along with its style configuration.

style: :construction: Design Update

Added new total style configuration to ETF, also added some image arrangements in mycard.html

feat: :sparkles: footer of balance

The configuration of the balance footer was added, with its editing in CSS

chore: :construction: general details

The old data was updated with the final data in index.html, in the same way the mycard info was styled, giving it the final position and generally colors and margins were modified

fix: :bug: update on mycard

The mycard credit cards were fixed from scratch, along with their configuration, and a code order was made in the same area, where space was cleaned.

chore: :lipstick: font configuration

The fonts were linked to each text in general of the virtualwallet and the font sizes were configured

fix: :art: general arrangements

general arrangements of margin, padding and border radius, also in balance.html an arrangement of images was made