release: :bookmark: version estable balance
creation of the folders that will be used during the project

feat: :tada: balance_html

feat: :sparkles: footer of balance

The configuration of the balance footer was added, with its editing in CSS

feat: :sparkles:  Balance update

The part of the balance value was added and arranged centrally, CSS configuration of the header of this same section was also added

feat: :construction: balance buttons

The buttons in the balance section were created and modified

feat: :construction: insertion of balance sheets

the balance sheet table was introduced

feat: :sparkles: spam entry

The balance spam information was entered and the balance days were also configured.

fix: :hammer: page layout update

configuration is modified and added in balance CSS so that the page begins to be responsive, with which many size and position configurations had to be refactored throughout the page

chore: :hammer: general configuration

The redirection of the pages was configured, as well as the types of colors and margins in balances and index were modified.

fix: :bug: update on mycard

The mycard credit cards were fixed from scratch, along with their configuration, and a code order was made in the same area, where space was cleaned.

fix: :art: general arrangements

general arrangements of margin, padding and border radius, also in balance.html an arrangement of images was made

feat: :fire: copia_balances_css

creation of the folders that will be used during the project

feat: :tada: balances_css

feat: :sparkles: footer of balance

The configuration of the balance footer was added, with its editing in CSS

feat: :sparkles:  Balance update

The part of the balance value was added and arranged centrally, CSS configuration of the header of this same section was also added

feat: :construction: balance buttons

The buttons in the balance section were created and modified

feat: :construction: insertion of balance sheets

the balance sheet table was introduced

feat: :sparkles: spam entry

The balance spam information was entered and the balance days were also configured.

fix: :hammer: page layout update

configuration is modified and added in balance CSS so that the page begins to be responsive, with which many size and position configurations had to be refactored throughout the page

fix: :lipstick: virtualWallet responsive

Virtual Wallet arrangement, responsive configuration for all pages

chore: :hammer: general configuration

The redirection of the pages was configured, as well as the types of colors and margins in balances and index were modified.

chore: :construction: general details

The old data was updated with the final data in index.html, in the same way the mycard info was styled, giving it the final position and generally colors and margins were modified

fix: :bug: update on mycard

The mycard credit cards were fixed from scratch, along with their configuration, and a code order was made in the same area, where space was cleaned.

chore: :lipstick: font configuration

The fonts were linked to each text in general of the virtualwallet and the font sizes were configured

fix: :art: general arrangements

general arrangements of margin, padding and border radius, also in balance.html an arrangement of images was made